+++ CHANGES.md
@@ -1,2 +1,10 @@
+## 0.1.0
++ add new fields to user profile
++ create Advisory Committee (AC) Members page
++ remove non-AC profiles in page template
+
+## 0.0.2
++ add title to login screen
+
 ## 0.0.1
 + getting started